Move the api inline
